modified:   ExcelCommons install.sql plsql/ExcelGen.pkb plsql/ExcelGen.pks
Add setNumFormat public procedure and associated functionality in the package
body. Operates the same way setDateFormat does. Not ideal in that it is for all
number columns and would prefer to specify it on a per column basis, but
that is beyond my kin and appears as if would take considerable redesign. This
works well enough for the most common use cases.

install.sql had a directory name change out from under it in one of the
recursive submodules.