Add set num format (#4)
* modified:   install.sql README.md plsql/ExcelGen.pkb plsql/ExcelGen.pks samples

Add setNumFormat public procedure and associated functionality in the package
body. Operates the same way setDateFormat does. Not ideal in that it is for all
number columns and would prefer to specify it on a per column basis, but
that is beyond my ken and appears as if would take considerable redesign. This
works well enough for the most common use cases.

add sample demonstrating setNumFormat

install.sql had a directory name change out from under it in one of the
recursive submodules. Fixed.

Update README.md with documentation for setNumFormat procedure

* Default Number Format left NULL. When no number format is specified,
leave as generic value.
Added test case function to validate default number format.

* correct check for default number format to look for numXfId=0

Co-authored-by: Lee Lindley <ltlindley@neiu.edu>